MODIFY: component 분리, 상수데이터 분리
