+++ src/pages/Login/AgreeInfo.js
@@ -1,0 +1,38 @@
+export const AGREE_INFO = [
+  {
+    id: 1,
+    title: '*이용 약관(필수)',
+    name: 'agree1',
+    term: '본 약관은 2017년 04월 06일부터 시행됩니다. 이용약관 Jo Malone London Korea Online 웹사이트(이하 본 사이트)에 오신 것을 환영합니다. Jo Malone London Korea Online(“이하 “회사”)는 아래의 약관, 개인정보처리방침 및 고객 서비스와 특정 기능, 게시물 또는 홍보와 관련하여 본 사이트 전체에서 제시하는 여타 조건과 정책 (본 약관(통칭하여, 본 약관이라 합니다)의 일부이며 그에 포함되는 것으로 간주됩니다)에 따라, 이용자에게 본 사이트상 컨텐츠와 서비스를 제공합니다. 이용자는 본 사이트에 접속하거나 이용함으로써 본 약관을 읽고, 이해하였으며, 일체의 제한요건이나 자격을 불문하고 본 약관에 따르기로 합의하였음을 인정하는 것입니다.',
+  },
+  {
+    id: 2,
+    title: '*개인정보의 수집 및 이용에 대한 동의(필수)',
+    name: 'agree2',
+    term: '회사는 본 웹사이트를 통한 회원가입 및 서비스 제공 등을 위하여 아래와 같은 개인정보를 수집하고 있습니다. - 이름, 비밀번호, 휴대폰 번호, 이메일 주소. 수집 및 이용 목적. 멤버십 계정 및 고객 데이터 관리, 회원제 서비스 이용에 따른 본인확인, 불량회원의 부정 이용 방지와 비인가 사용 방지, 가입 의사 확인, 연령확인, 불만처리 등 민원처리, 고지사항 전달',
+  },
+  {
+    id: 3,
+    title: '마케팅 수신 및 홍보 목적의 개인정보 수집 및 이용에 대한 동의(선택)',
+    name: 'agree3',
+    term: '수집하는 개인정보 항목 • 이름, 휴대폰 번호, 이메일 주소. 수집 및 이용 목적. SMS, DM, 이메일 등 전자적 전송 매체를 통한 개인 맞춤형 광고 전송, 이벤트 및 프로모션 관련 정보 안내, 신규 서비스(제품)개발 및 특화, 접속 빈도 파악 또는 회원의 서비스 이용에 대한 통계',
+  },
+];
+
+export const RECEPTION = [
+  {
+    id: 1,
+    title: '뉴스레터 수신여부',
+    name: 'reception1',
+  },
+  {
+    id: 2,
+    title: 'MMS 수신여부',
+    name: 'reception2',
+  },
+  {
+    id: 3,
+    title: 'DM 수신여부',
+    name: 'reception3',
+  },
+];
