+++ src/pages/Login/Login.scss
@@ -1,20 +1,3 @@
-@mixin buttonStyle {
-  width: 100%;
-  border: none;
-  background-color: black;
-  color: white;
-}
-
-@mixin inputStyle {
-  width: 100%;
-  height: 40px;
-  margin: 10px 0;
-  padding: 0 10px;
-  border: 1px solid #cccccc;
-  border-radius: 0%;
-  line-height: middle;
-}
-
 .login {
   background-color: #fdf9ef;
 
@@ -34,153 +17,6 @@
       display: flex;
       justify-content: space-between;
       width: 100%;
-
-      .newCustomer {
-        width: 45%;
-        margin: 50px 0;
-
-        .bold {
-          font-weight: bold;
-        }
-
-        .inputSignup input {
-          @include inputStyle;
-
-          &:focus {
-            outline: 1px dotted black;
-          }
-
-          &::placeholder {
-            color: #cccccc;
-          }
-        }
-
-        .pwCautionMsg {
-          margin: 20px 0;
-          color: #d80000;
-          font-size: 14px;
-          line-height: 25px;
-        }
-
-        .phoneNumber {
-          font-size: 14px;
-        }
-
-        .phoneNumInput {
-          select,
-          input {
-            width: 32%;
-            height: 40px;
-            margin-top: 5px;
-            margin-right: 5px;
-            padding: 0 10px;
-            border: 1px solid #cccccc;
-
-            &:focus {
-              outline: 1px dotted black;
-            }
-
-            &::placeholder {
-              color: #cccccc;
-            }
-          }
-        }
-
-        .signupBtn {
-          @include buttonStyle;
-          height: 40px;
-          margin-top: 8px;
-          margin-bottom: 13px;
-        }
-
-        .certificationNum {
-          width: 100%;
-          height: 40px;
-          margin: 10px 0;
-          padding: 0 10px;
-          border: 1px solid #cccccc;
-          border-radius: 0%;
-          line-height: middle;
-        }
-
-        .agreement {
-          .agreementBold {
-            margin-top: 15px;
-            font-size: 14px;
-            font-weight: bold;
-          }
-
-          .scrollBarBox {
-            width: 100%;
-            height: 150px;
-            margin: 10px 0;
-            padding: 10px;
-            overflow-y: scroll;
-            border: 1px solid #cccccc;
-            background-color: #fdf9ef;
-            font-size: 14px;
-            line-height: 25px;
-          }
-
-          .agree {
-            font-size: 14px;
-          }
-
-          .allAgree {
-            margin: 10px 0;
-            font-size: 14px;
-          }
-
-          .reception {
-            margin: 30px 0;
-            font-size: 14px;
-          }
-
-          .agreementMsg {
-            color: red;
-            font-size: 12px;
-            line-height: 25px;
-          }
-
-          .olderThanFourteen {
-            margin-top: 10px;
-            margin-bottom: 30px;
-            font-size: 14px;
-          }
-        }
-      }
-
-      .existingCustomers {
-        width: 50%;
-        margin: 50px 0;
-
-        .bold {
-          font-weight: bold;
-        }
-
-        .inputLogin input {
-          @include inputStyle;
-
-          &:focus {
-            outline: 1px dotted black;
-          }
-
-          &::placeholder {
-            color: #cccccc;
-          }
-        }
-
-        .underline {
-          font-size: 12px;
-          text-decoration: underline;
-        }
-
-        .loginBtn {
-          @include buttonStyle;
-          height: 38px;
-          margin-top: 40px;
-        }
-      }
     }
   }
 }
