+++ src/pages/Login/styles/mixin.scss
@@ -1,0 +1,16 @@
+@mixin buttonStyle {
+  width: 100%;
+  border: none;
+  background-color: black;
+  color: white;
+}
+
+@mixin inputStyle {
+  width: 100%;
+  height: 40px;
+  margin: 10px 0;
+  padding: 0 10px;
+  border: 1px solid #cccccc;
+  border-radius: 0%;
+  line-height: middle;
+}
